changed schema for bigint and cents; added conversion helpers

+++ migrations/001.do.create_users_table.sql
@@ -6,6 +6,6 @@ CREATE TABLE IF NOT EXISTS "users" (
   "date_created" TIMESTAMPTZ DEFAULT now() NOT NULL,
   "password" TEXT NOT NULL,
   "email" TEXT NOT NULL,
-  "allowance" NUMERIC(12,2) CHECK( allowance >= 0) DEFAULT 0 NOT NULL ,
-  "balance" NUMERIC(12,2) DEFAULT 0 NOT NULL 
+  "allowance" BIGINT CHECK(allowance >= 0) DEFAULT 0 NOT NULL ,
+  "balance" BIGINT DEFAULT 0 NOT NULL 
 );
